+++ backend/tests/Unit/Service/PluginService/plugin-both/src/TestBoth.php
@@ -7,6 +7,7 @@ declare(strict_types=1);
 namespace Tests\Unit\Service\PluginService\plugin\src;
 
 use Neucore\Plugin\Core\FactoryInterface;
+use Neucore\Plugin\Core\OutputInterface;
 use Neucore\Plugin\Data\CoreAccount;
 use Neucore\Plugin\Data\CoreCharacter;
 use Neucore\Plugin\Data\PluginConfiguration;
@@ -43,6 +44,10 @@ class TestBoth implements GeneralInterface, ServiceInterface
     public function getNavigationItems(): array
     {
         return [];
+    }
+
+    public function command(array $arguments, array $options, OutputInterface $output): void
+    {
     }
 
     public function getAccounts(array $characters): array